chore: add tmp cli mod
